Use network_get to retrieve the IP of the correct reference interface
As per bug https://github.com/openstack-charmers/magpie-layer/issues/17,
the use of the hook tool `unit-get private-address` is problematic since
it doesn't always get the correct IP when there are multiple interfaces
on the machine. This change ensures that we explicitly state that we
want to get the IP of the magpie binding.