Add support for min_speed percentage of link speed
This can be useful for environments with mixed link speeds.

Example:

```
juju config magpie min_speed=80%
```

Further note on behaviour: if a percentage min_speed is set,
and the link is virtual or otherwise returns a negative speed,
magpie will fail the speed check.
(percentage of an undefined value is still undefined)

Closes-Bug: #1977784

Change-Id: I1fa5d38a4da26ca38e98350d1a1625184dff2363