Optionally ignore LACP activity bit in port check
If the LACP bond on the switch side is configured in passive mode,
the activity mode bit on the port states will usually be mismatched.

Add an option to the charm config to inform when this is expected,
so that the charm will ignore this mismatched bit when checking ports.

Closes-Bug: #1958327

Change-Id: I784bb410110813472e14b3477247fe138db5c214