Display absolute stats on ping failures
Previously the percentage of packet loss was displayed,
but we received feedback that more detail was required.
So we update the output to include the total number of pings
and the number of successful pings.

Example of how this looks in the status line:

magpie/0*  blocked   idle   0        10.62.174.150          icmp failed: 1: 0/20 packets received, local hostname ok (juju-f0966f-0), dns ok, iperf leader, mtu: 1500

Change-Id: I8bf49bed8e8c8b7dd8497139a0b613650848e98b